requirements.txt
flask>=2.1.0,<3.0.0
flask-cors>=3.0.10,<4.0.0
Flask-SQLAlchemy>=2.5.1,<3.0.0
flask-login>=0.6.2,<0.7.0
flask-migrate>=3.1.0,<5.0.0
requests>=2.26.0,<3.0.0
openai>=1.0.0,<2.0.0
replicate>=0.15.0,<1.0.0
sendgrid>=6.10.0,<7.0.0
python-wordpress-xmlrpc@git+https://github.com/maxcutler/python-wordpress-xmlrpc.git
python-dotenv>=0.19.2,<2.0.0
pymysql>=1.0.2,<2.0.0
pytz>=2021.3
Werkzeug>=2.0.3,<3.0.0
SQLAlchemy<2.0
APScheduler==3.10.1